the meat & potatoes
There are two important methods at work here, writeSettings() and setSettings(). Write settings would be called when the player saves settings he set in the theoretical game options menu, alternatively writeSettings() is used to generate a default config.properties file if setSettings() throws a FileNotFoundException. The setSettings() method would be called once when the game is being initialized in order to set game engine attributes. If the config.properties file exists but contains values outside the accepted characters (in this case integers of 1-5) a default value is used instead.